requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
python-dotenv
twilio
stripe
passlib[bcrypt]
pyjwt

